Added Drag Drop User Interface
Cleaner looking then what we had in the original fits law file. The
numbered items in the top can be dragged to the Shopping bag area. If
the item is not dragged to the Shopping bag then it goes back to the
top. My next step is to track the mouse movement and store it to an
object in order to do calculations.